% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{check_python_and_module}
\alias{check_python_and_module}
\title{Check Python and Required Modules Availability}
\usage{
check_python_and_module(module = "scipy", initialize = TRUE)
}
\arguments{
\item{module}{Character. Name of the Python module to check (default is "scipy").}

\item{initialize}{Logical. Whether to initialize Python if not already initialized (default is TRUE).}
}
\value{
This function does not return a value. It stops execution if the required Python environment is not available.
}
\description{
This helper function checks if Python is available and if the required Python module (e.g., \code{scipy}) is installed.
It stops execution and raises an error if Python or the specified module is not available.
}
\examples{
\dontrun{
check_python_and_module("scipy") # Check for Python and 'scipy'
}

}
