% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{create_package_manifest}
\alias{create_package_manifest}
\title{Function to Create MANIFEST.txt}
\usage{
create_package_manifest(paths, manifest_path = "MANIFEST.txt", temp_dir)
}
\arguments{
\item{paths}{A character vector of paths to files and/or directories to include in the manifest.}

\item{manifest_path}{A character string specifying the path to the manifest file. Default is "MANIFEST.txt".}

\item{temp_dir}{A character string specifying the temporary directory to be removed from the file paths.}
}
\description{
This function generates a MANIFEST.txt file that lists all files in the specified paths,
along with their sizes. It recursively includes files from directories and skips paths that
do not exist. The manifest excludes the manifest file itself if present in the list.
}
