% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{extract_aphia_id}
\alias{extract_aphia_id}
\title{Extract the AphiaID from the First Row of Each worms_records Tibble}
\usage{
extract_aphia_id(record)
}
\arguments{
\item{record}{A tibble containing worms_records with at least an 'AphiaID' column.}
}
\value{
A numeric value representing the AphiaID of the first row in the tibble,
or NA if the tibble is empty.
}
\description{
This function extracts the AphiaID from the first row of a given worms_records tibble.
If the tibble is empty, it returns NA.
}
\examples{
# Example usage:
record <- dplyr::tibble(AphiaID = c(12345, 67890))
iRfcb:::extract_aphia_id(record)

empty_record <- dplyr::tibble(AphiaID = numeric(0))
iRfcb:::extract_aphia_id(empty_record)
}
