% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{extract_features}
\alias{extract_features}
\title{Extract Features and Add Sample Names}
\usage{
extract_features(sample_name, feature_data)
}
\arguments{
\item{sample_name}{Character. The name of the sample, typically derived from the feature file name.}

\item{feature_data}{Dataframe. The feature data associated with the sample.}
}
\value{
A dataframe with the sample name added as a column.
}
\description{
This helper function adds the sample name as a column in the feature dataframe.
It assumes that the sample name is embedded in the filename, and this function strips unnecessary parts of the filename.
}
\examples{
\dontrun{
sample_name <- "sample_001_fea_v2.csv"
feature_data <- data.frame(roi_number = 1:10, feature_value = rnorm(10))
result <- extract_features(sample_name, feature_data)
}
}
