% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxa_images_from_ROI.R
\name{extract_images_from_ROI}
\alias{extract_images_from_ROI}
\title{Extract Images from ROI File}
\usage{
extract_images_from_ROI(
  roifile,
  outdir = dirname(roifile),
  taxaname = NULL,
  ROInumbers = NULL
)
}
\arguments{
\item{roifile}{A character string specifying the path to the .roi file.}

\item{outdir}{A character string specifying the directory where the PNG images will be saved. Defaults to the directory of the ROI file.}

\item{taxaname}{An optional character string specifying the taxa name for the subdirectory where images will be saved. Defaults to NULL.}

\item{ROInumbers}{An optional numeric vector specifying the ROI numbers to extract. If NULL, all ROIs with valid dimensions are extracted.}
}
\value{
No return value, called for side effects. Writes PNG images to a directory.
}
\description{
This function reads a .roi file and its corresponding .adc file, extracts specified regions of interest (ROIs),
and saves each ROI as a PNG image in a specified directory. Optionally, you can specify a taxa name and ROI numbers.
}
\examples{
\dontrun{
# Convert ROI file to PNG images
extract_images_from_ROI("your_roi_file.roi")

# Extract taxa images from ROI file
extract_images_from_ROI("your_roi_file.roi", "output_directory", "taxa_name")
}
}
