% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{extract_parts}
\alias{extract_parts}
\title{Function to extract parts using regular expressions}
\usage{
extract_parts(filename)
}
\arguments{
\item{filename}{A character string specifying the filename to extract parts from.}
}
\value{
A data frame with columns: sample, timestamp, date, year, month, day, time, and ifcb_number.
}
\description{
This function extracts timestamp, IFCB number, and date components from a filename.
}
