% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_taxa_images.R
\name{extract_taxa_images}
\alias{extract_taxa_images}
\title{Extract Taxa Images from Classified Sample}
\usage{
extract_taxa_images(
  sample,
  classifier,
  ifcb_path,
  taxa = "All",
  ifcb_unit = "IFCB134",
  version = "v1"
)
}
\arguments{
\item{sample}{A character string specifying the sample name.}

\item{classifier}{A character string specifying the classifier name.}

\item{ifcb_path}{A character string specifying the base path to the IFCB data.}

\item{taxa}{A character string specifying the taxa to extract. Default is "All".}

\item{ifcb_unit}{A character string specifying the IFCB unit. Default is "IFCB134".}

\item{version}{A character string specifying the version of the classified data. Default is "v1".}
}
\value{
No return value, called for side effects. Extracts and saves taxa images to a directory.
}
\description{
This function reads a classified sample file, extracts specified taxa images from the corresponding ROI files,
and saves each image in a specified directory.
}
\examples{
\dontrun{
# Extract taxa images from classified sample
classifier <- "Baltic"
sample <- "D20230311T092911"
extract_taxa_images(sample, classifier, "path/to/ifcb_data")
}
}
