% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{find_matching_data}
\alias{find_matching_data}
\title{Function to Find Matching Data Files with a General Pattern}
\usage{
find_matching_data(mat_file, data_files)
}
\arguments{
\item{mat_file}{A character string specifying the path to the .mat file.}

\item{data_files}{A character vector of paths to data files to search.}
}
\value{
A character vector of matching data files.
}
\description{
This function finds data files that match the base name of a given .mat file.
}
