% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{find_matching_features}
\alias{find_matching_features}
\title{Function to Find Matching Feature Files with a General Pattern}
\usage{
find_matching_features(mat_file, feature_files)
}
\arguments{
\item{mat_file}{A character string specifying the path to the .mat file.}

\item{feature_files}{A character vector of paths to feature files to search.}
}
\value{
A character vector of matching feature files.
}
\description{
This function finds feature files that match the base name of a given .mat file.
}
