% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class2use.R
\name{get_class2use}
\alias{get_class2use}
\title{Get Classes from Class2use File}
\usage{
get_class2use(class2use_file)
}
\arguments{
\item{class2use_file}{A character string specifying the path to the .mat file containing class information.}
}
\value{
A character vector of class names.
}
\description{
This function reads a .mat file containing class information and extracts the list of classes.
}
\examples{
\dontrun{
# Get class names from a class2use file
classes <- get_class2use("path/to/class2use.mat")
print(classes)
}
}
