% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_class2use_from_classifer.R
\name{get_class2use_from_classifer}
\alias{get_class2use_from_classifer}
\title{Get Classes from Classifier File}
\usage{
get_class2use_from_classifer(class_file)
}
\arguments{
\item{class_file}{A character string specifying the path to the .mat file containing class information.}
}
\value{
A character vector of class names.
}
\description{
This function reads a .mat file containing class information and extracts the list of classes.
}
\examples{
\dontrun{
# Get class names from a classifier file
classes <- get_class2use_from_classifer("path/to/classifier.mat")
print(classes)
}
}
