% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classes_from_mat.R
\name{get_classes_from_mat}
\alias{get_classes_from_mat}
\title{Get Classes from a MAT File}
\usage{
get_classes_from_mat(mat_file, variable_name = "class2use")
}
\arguments{
\item{mat_file}{A character string specifying the path to the .mat file containing class information.}

\item{variable_name}{A character string specifying the variable name in the .mat file that contains the class information.
Default is "class2use".}
}
\value{
A character vector of class names.
}
\description{
This function reads a .mat file containing class information and extracts the list of classes.
}
\examples{
\dontrun{
# Get class names from a class2use file
classes <- get_classes_from_mat("path/to/class2use.mat", "class2use")
print(classes)

# Get class names from a classifier file
classes <- get_classes_from_mat("path/to/classifier.mat", "classes")
print(classes)
}
}
