% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_names_from_mat.R
\name{get_variable_names_from_mat}
\alias{get_variable_names_from_mat}
\title{Get Variable Names from a MAT File}
\usage{
get_variable_names_from_mat(mat_file)
}
\arguments{
\item{mat_file}{A character string specifying the path to the .mat file.}
}
\value{
A character vector of variable names.
}
\description{
This function reads a .mat file and retrieves the names of all variables stored within it.
}
\examples{
\dontrun{
# Get variable names from a MAT file
variables <- get_variable_names_from_mat("path/to/file.mat")
print(variables)
}
}
\seealso{
\code{\link{get_classes_from_mat}}
}
