% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{handle_missing_ferrybox_data}
\alias{handle_missing_ferrybox_data}
\title{Handle Missing Ferrybox Data}
\usage{
handle_missing_ferrybox_data(
  data,
  ferrybox_data,
  parameters,
  rounding_function
)
}
\arguments{
\item{data}{A data frame containing the main dataset with a \code{timestamp} column and several parameter columns that might have missing data.}

\item{ferrybox_data}{A data frame containing the Ferrybox dataset. This dataset should have a \code{timestamp} column and columns corresponding to the parameters specified.}

\item{parameters}{A character vector of column names (parameters) in \code{data} that should be checked for missing values and potentially filled using the Ferrybox data.}

\item{rounding_function}{A function that rounds the \code{timestamp} to a specified unit (e.g., minute). This function should take a \code{timestamp} column and a \code{unit} argument.}
}
\value{
A data frame similar to \code{data}, but with missing values in the specified parameters filled in using the Ferrybox data. The output includes only the \code{timestamp} and the specified parameter columns.
}
\description{
This function processes a dataset by filling in missing values for specific parameters using data from a Ferrybox dataset. It rounds the timestamp to a specified unit, joins the Ferrybox data based on the rounded timestamp, and fills in missing values using the corresponding Ferrybox data.
}
\details{
The function performs the following steps:
\itemize{
\item Renames the columns in the \code{ferrybox_data} by appending \verb{_ferrybox} to the names of the specified parameters.
\item Filters the \code{data} for rows with missing values in any of the specified parameters.
\item Rounds the \code{timestamp} to the nearest specified unit using the \code{rounding_function}.
\item Joins the \code{ferrybox_data} to the main \code{data} based on the rounded \code{timestamp}.
\item Uses the \code{coalesce} function to fill in missing values in the specified parameters with the corresponding values from the Ferrybox data.
\item Returns a cleaned dataset containing only the \code{timestamp} and the specified parameter columns.
}
}
\examples{
# Assuming you have a data frame `data` with missing values, a Ferrybox data frame `ferrybox_data`,
# and a rounding function `round_timestamp`.
filled_data <- handle_missing_ferrybox_data(data, ferrybox_data, c("8002", "8003", "8172"), round_timestamp)

}
