% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_get_mat_classes.R
\name{ifcb_get_mat_classes}
\alias{ifcb_get_mat_classes}
\title{Get Classes from a MAT File}
\usage{
ifcb_get_mat_classes(mat_file, variable_name = "class2use")
}
\arguments{
\item{mat_file}{A character string specifying the path to the .mat file containing class information.}

\item{variable_name}{A character string specifying the variable name in the .mat file that contains the class information.
Default is "class2use". Other examples are class2use.manual from a manual file, or class2use.auto for an
classlist used for automatic assignment. This parameter can be found using `get_variable_names_from_mat`.}
}
\value{
A character vector of class names.
}
\description{
This function reads a .mat file generated the ifcb-analysis repository,
containing class information and extracts the list of classes.
}
\examples{
\dontrun{
# Get class names from a class2use file
classes <- ifcb_get_mat_classes("path/to/class2use.mat", "class2use")
print(classes)

# Get class names from a classifier file
classes <- ifcb_get_mat_classes("path/to/classifier.mat", "classes")
print(classes)
}
}
\seealso{
\code{\link{get_variable_names_from_mat}} \url{https://github.com/hsosik/ifcb-analysis}
}
