% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_get_mat_variables.R
\name{ifcb_get_mat_variables}
\alias{ifcb_get_mat_variables}
\title{Get Variable Names from a MAT File}
\usage{
ifcb_get_mat_variables(mat_file)
}
\arguments{
\item{mat_file}{A character string specifying the path to the .mat file.}
}
\value{
A character vector of variable names.
}
\description{
This function reads a .mat file generated the ifcb-analysis repository and retrieves the
names of all variables stored within it.
}
\examples{
\dontrun{
# Get variable names from a MAT file
variables <- ifcb_get_mat_variables("path/to/file.mat")
print(variables)
}
}
\seealso{
\code{\link{ifcb_get_mat_classes}}  \url{https://github.com/hsosik/ifcb-analysis}
}
