% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_get_trophic_type.R
\name{ifcb_get_trophic_type}
\alias{ifcb_get_trophic_type}
\title{Get Trophic Type for a List of Taxa}
\usage{
ifcb_get_trophic_type(taxa_list)
}
\arguments{
\item{taxa_list}{A character vector of scientific names for which trophic types are to be retrieved.}
}
\value{
A character vector of trophic types corresponding to the scientific names in \code{taxa_list}.
}
\description{
This function reads a file containing trophic types for various taxa (originating from SMHI),
             and returns the trophic type for a specified list of taxa.
             If there are multiple unique trophic types for a scientific name,
             the summarized trophic type is "NS".
}
\examples{
# Example usage:
taxa_list <- c("Acanthoceras zachariasii",
               "Nodularia spumigena",
               "Acanthoica quattrospina",
               "Noctiluca",
               "Gymnodiniales")

ifcb_get_trophic_type(taxa_list)

}
