% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_is_in_baltic.R
\name{ifcb_is_in_baltic}
\alias{ifcb_is_in_baltic}
\title{Check if Points are in the Baltic Sea Basin}
\usage{
ifcb_is_in_baltic(latitudes, longitudes)
}
\arguments{
\item{latitudes}{A numeric vector of latitude points.}

\item{longitudes}{A numeric vector of longitude points.}
}
\value{
A logical vector indicating whether each point is in the Baltic Sea Basin.
}
\description{
This function checks if vectors of latitude and longitude points are within the
Baltic Sea (including Öresund) using a pre-packaged shapefile included in the `iRfcb` package.
}
\details{
This function extracts a pre-packaged shapefile of the Baltic Sea Basin from the `iRfcb` package, sets the CRS,
transforms the CRS to WGS84 (EPSG:4326), and checks if the given points fall within the Baltic Sea Basin.
}
\examples{
# Define example latitude and longitude vectors
latitudes <- c(55.337, 54.729, 56.311, 57.975)
longitudes <- c(12.674, 14.643, 12.237, 10.637)

# Check if the points are in the Baltic Sea Basin
result <- ifcb_is_in_baltic(latitudes, longitudes)
print(result)

}
