% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_psd_plot.R
\name{ifcb_psd_plot}
\alias{ifcb_psd_plot}
\title{Generate PSD Plot for a Given Sample}
\usage{
ifcb_psd_plot(sample_name, data, fits, start_fit)
}
\arguments{
\item{sample_name}{The name of the sample to plot.}

\item{data}{A data frame containing the PSD data, where each row represents a sample and each column represents the PSD values at different particle sizes.}

\item{fits}{A data frame containing the fit parameters for the power curve, where each row represents a sample and the columns include the parameters `a`, `k`, and `R2`.}

\item{start_fit}{The x-value threshold below which data should be excluded from the plot and fit.}
}
\value{
A ggplot object representing the PSD plot for the sample.
}
\description{
This function generates a plot for a given sample from PSD data and fits.
}
\examples{
\dontrun{
plot <- ifcb_psd_plot(sample_name = "Sample1", data = data, fits = fits, start_fit = 10)
}
}
