% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_run_image_gallery.R
\name{ifcb_run_image_gallery}
\alias{ifcb_run_image_gallery}
\title{Run IFCB Image Gallery}
\usage{
ifcb_run_image_gallery()
}
\value{
No return value. This function launches a Shiny application for interactive image browsing and management.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated as there is a better alternative: \code{ClassiPyR::run_app()}. For more information,
please see: \url{https://europeanifcbgroup.github.io/ClassiPyR/}

Launches a Shiny application that provides an interactive interface for
browsing and managing IFCB (Imaging FlowCytobot) image galleries.

Users can specify a folder containing .png images, navigate through the images,
select and unselect images, and download a list of selected images.
This feature is particularly useful for quality control of annotated images.
A downloaded list of images from the app can also be uploaded to filter and
view only the selected images.
}
\examples{
\donttest{
# Run the IFCB image gallery Shiny app
if(interactive()){
  ifcb_run_image_gallery()
}
}
}
