% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_summarize_class_counts.R
\name{ifcb_summarize_class_counts}
\alias{ifcb_summarize_class_counts}
\title{Count Cells from TreeBagger Classifier Output}
\usage{
ifcb_summarize_class_counts(classpath_generic, hdr_folder, year_range)
}
\arguments{
\item{classpath_generic}{Character string specifying the location of the classifier output files.
The path should include 'xxxx' in place of the 4-digit year (e.g., 'classxxxx_v1/').}

\item{hdr_folder}{Character string specifying the directory where the data (hdr files) are located.
This can be a URL for web services or a full path for local files.}

\item{year_range}{Numeric vector specifying the range of years (e.g., 2013:2014) to process.}
}
\value{
A list containing the following elements:
  \item{class2useTB}{Classes used in the TreeBagger classifier.}
  \item{classcountTB}{Counts of each class considering each target placed in the winning class.}
  \item{classcountTB_above_optthresh}{Counts of each class considering only classifications above the optimal threshold for maximum accuracy.}
  \item{ml_analyzedTB}{Volume analyzed for each file.}
  \item{mdateTB}{Dates associated with each file.}
  \item{filelistTB}{List of files processed.}
  \item{classpath_generic}{The generic classpath provided as input.}
  \item{classcountTB_above_adhocthresh (optional)}{Counts of each class considering only classifications above the adhoc threshold.}
  \item{adhocthresh (optional)}{The adhoc threshold used for classification.}
}
\description{
This function summarizes class results for a series of classifier output files
and returns a summary data list.
}
\examples{
\dontrun{
ifcb_summarize_class_counts('C:/work/IFCB/user_training_test_data/class/classxxxx_v1/',
                            'C:/work/IFCB/user_training_test_data/data/', 2014)
}

}
