% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_zip_pngs.R
\name{ifcb_zip_pngs}
\alias{ifcb_zip_pngs}
\title{Zip PNG Folders}
\usage{
ifcb_zip_pngs(
  png_folder,
  zip_filename,
  readme_file = NULL,
  email_address = "",
  version = ""
)
}
\arguments{
\item{png_folder}{The directory containing subdirectories with .png files.}

\item{zip_filename}{The name of the zip file to create.}

\item{readme_file}{Optional path to a README file for inclusion in the zip package.}

\item{email_address}{Optional email address to include in the README file.}

\item{version}{Optional version information to include in the README file.}
}
\value{
This function does not return any value; it creates a zip archive and potentially a README file.
}
\description{
This function zips directories containing .png files and optionally includes README and MANIFEST files.
}
\examples{
# Zip all subdirectories in the 'images' folder with a README file
ifcb_zip_pngs("path/to/images",
             "images.zip",
             readme_file = system.file("inst/exdata/README-template.md", package = "iRfcb"),
             email_address = "example@example.com",
             version = "1.0")

# Zip all subdirectories in the 'images' folder without a README file
ifcb_zip_pngs("path/to/images", "images.zip")

}
\seealso{
\code{\link{ifcb_zip_matlab}}
}
