% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{print_progress}
\alias{print_progress}
\title{Function to Print the Progress Bar}
\usage{
print_progress(current, total, bar_width = 50)
}
\arguments{
\item{current}{An integer specifying the current progress.}

\item{total}{An integer specifying the total steps for the process.}

\item{bar_width}{An integer specifying the width of the progress bar. Default is 50.}
}
\description{
This function prints a progress bar to the console to indicate the progress of a process.
}
