% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_value_in_classlist.R
\name{replace_value_in_classlist}
\alias{replace_value_in_classlist}
\title{Replace Specific Values in Classlist}
\usage{
replace_value_in_classlist(
  input_file,
  output_file,
  target_value,
  new_value,
  column_index = 2
)
}
\arguments{
\item{input_file}{A character string specifying the path to the input .mat file.}

\item{output_file}{A character string specifying the path to the output .mat file.}

\item{target_value}{The value to be replaced in the specified column of the classlist.}

\item{new_value}{The new value to replace the target value with.}

\item{column_index}{A numeric value specifying the index of the column where the replacement should occur (1-based index). Defaults to 2.}
}
\value{
No return value, called for side effects. Writes modified .mat file to the specified directory.
}
\description{
This function loads a .mat file, replaces specified target values in a specific column of the 'classlist' field with a new value, and saves the modified content to a new .mat file.
}
\examples{
\dontrun{
# Replace specific values in the classlist column
replace_value_in_classlist("path/to/input_file.mat", "path/to/output_file.mat", "target_value", "new_value", 2)
}
}
