% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{retrieve_worms_records}
\alias{retrieve_worms_records}
\title{Retrieve WoRMS Records with Retry Mechanism}
\usage{
retrieve_worms_records(taxa_names, max_retries = 3, sleep_time = 10)
}
\arguments{
\item{taxa_names}{A character vector of taxa names to retrieve records for.}

\item{max_retries}{An integer specifying the maximum number of attempts to retrieve records.}

\item{sleep_time}{A numeric value indicating the number of seconds to wait between retry attempts.}
}
\value{
A list of WoRMS records or NULL if the retrieval fails after the maximum number of attempts.
}
\description{
This helper function attempts to retrieve WoRMS records using the provided taxa names.
It retries the operation if an error occurs, up to a specified number of attempts.
}
