% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_images_by_class.R
\name{summarize_images_by_class}
\alias{summarize_images_by_class}
\title{Summarize Image Counts by Class and Sample}
\usage{
summarize_images_by_class(main_directory, hdr_directory = NULL, verbose = TRUE)
}
\arguments{
\item{main_directory}{A character string specifying the path to the main directory containing subfolders (classes) with .png images.}

\item{hdr_directory}{A character string specifying the path to the directory containing the .hdr files. Default is NULL.}

\item{verbose}{A logical indicating whether to print progress messages. Default is TRUE.}
}
\value{
A data frame with columns: sample, ifcb_number, class_name, n_images, gpsLatitude, gpsLongitude, timestamp, year, month, day, time, roi_numbers.
}
\description{
This function summarizes the number of images per class for each sample,
and optionally retrieves GPS positions, timestamps, and IFCB information using extract_hdr_data and convert_ifcb_filenames functions.
}
