% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{truncate_folder_name}
\alias{truncate_folder_name}
\title{Function to Truncate the Folder Name}
\usage{
truncate_folder_name(folder_name)
}
\arguments{
\item{folder_name}{A character string specifying the folder name to truncate.}
}
\value{
A character string with the truncated folder name.
}
\description{
This function removes the trailing underscore and three digits from the base name of a folder.
}
