% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_png_folders.R
\name{zip_png_folders}
\alias{zip_png_folders}
\title{Zip PNG Folders}
\usage{
zip_png_folders(
  png_directory,
  zip_filename,
  readme_file = NULL,
  email_address = "",
  version = ""
)
}
\arguments{
\item{png_directory}{The directory containing subdirectories with .png files.}

\item{zip_filename}{The name of the zip file to create.}

\item{readme_file}{Optional path to a README file for inclusion in the zip package.}

\item{email_address}{Optional email address to include in the README file.}

\item{version}{Optional version information to include in the README file.}
}
\value{
This function does not return any value; it creates a zip archive and potentially a README file.
}
\description{
This function zips directories containing .png files and optionally includes a README file.
}
\examples{
# Zip all subdirectories in the 'images' folder with a README file
zip_png_folders("path/to/images", "images.zip", readme_file = "path/to/README.md", email_address = "example@example.com", version = "1.0")

# Zip all subdirectories in the 'images' folder without a README file
zip_png_folders("path/to/images", "images.zip")

}
