% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_count_mat_annotations.R
\name{ifcb_count_mat_annotations}
\alias{ifcb_count_mat_annotations}
\title{Count IFCB Annotations from .mat Files}
\usage{
ifcb_count_mat_annotations(
  manual_folder,
  class2use_file,
  skip_class = NULL,
  sum_level = "class"
)
}
\arguments{
\item{manual_folder}{A character string specifying the path to the folder containing .mat files.}

\item{class2use_file}{A character string specifying the path to the file containing the class2use variable.}

\item{skip_class}{A numeric vector of class IDs or a character vector of class names to be excluded from the count. Default is NULL.}

\item{sum_level}{A character string specifying the level of summarization. Options: "sample" or "class" (default).}
}
\value{
A data frame with the total count of images per class or per sample.
}
\description{
This function processes .mat files, generated by the code in the `ifcb-analysis` repository (Sosik and Olson 2007),
in a specified folder to count and summarize the annotations for each class based on the class2use information provided in a file.
}
\examples{
\dontrun{
# Count annotations excluding specific class IDs
result <- ifcb_count_mat_annotations("path/to/manual_folder",
                                     "path/to/class2use_file",
                                     skip_class = c(99, 100))
print(result)

# Count annotations excluding a specific class name
result <- ifcb_count_mat_annotations("path/to/manual_folder",
                                     "path/to/class2use_file",
                                     skip_class = "unclassified")
print(result)
}
}
\references{
Sosik, H. M. and Olson, R. J. (2007), Automated taxonomic classification of phytoplankton sampled with imaging-in-flow cytometry. Limnol. Oceanogr: Methods 5, 204–216.
}
