% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_download_dashboard_data.R
\name{ifcb_download_dashboard_data}
\alias{ifcb_download_dashboard_data}
\title{Download IFCB data files from an IFCB Dashboard}
\usage{
ifcb_download_dashboard_data(
  dashboard_url,
  samples,
  file_types,
  dest_dir,
  convert_filenames = FALSE,
  convert_adc = FALSE,
  parallel_downloads = 10,
  sleep_time = 2,
  multi_timeout = 120,
  quiet = FALSE
)
}
\arguments{
\item{dashboard_url}{Character. The base URL of the IFCB dashboard (e.g., \code{"https://ifcb-data.whoi.edu"}).
If no subpath (e.g., \verb{/data/} or \verb{/mvco/}) is included, \verb{/data/} will be added automatically.}

\item{samples}{Character vector. The IFCB sample identifiers (e.g., \code{"IFCB1_2014_188_222013"} or \code{"D20220807T025424_IFCB010"}).}

\item{file_types}{Character vector. Specifies which file types to download.
Allowed values: \code{"blobs"}, \code{"features"}, \code{"autoclass"}, \code{"roi"}, \code{"zip"}, \code{"hdr"}, \code{"adc"}.}

\item{dest_dir}{Character. The directory where downloaded files will be saved.}

\item{convert_filenames}{Logical. If \code{TRUE}, converts filenames of the old format \code{"IFCBxxx_YYYY_DDD_HHMMSS"}
to the new format (\code{DYYYYMMDDTHHMMSS_IFCBXXX}).
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}

\item{convert_adc}{Logical. If \code{TRUE}, adjusts ADC files from older IFCB instruments
(IFCB1–6, with filenames in the format \code{"IFCBxxx_YYYY_DDD_HHMMSS"}) by inserting
four empty columns after column 7 to match the newer format. Default is \code{FALSE}.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}

\item{parallel_downloads}{Integer. The number of files to download in parallel per batch.
This helps manage network load and system performance. Default is \code{10}.}

\item{sleep_time}{A numeric value indicating the number of seconds to wait between each batch of downloads. Default is \code{2}.}

\item{multi_timeout}{Numeric. The maximum time in seconds that the \code{curl} multi-download request
will wait for a response before timing out. This helps prevent
hanging downloads in case of slow or unresponsive servers. Default is \code{120} seconds.}

\item{quiet}{Logical. If TRUE, suppresses messages about the progress and completion of the download process. Default is FALSE.}
}
\value{
This function does not return a value. It performs the following actions:
\itemize{
\item Downloads the requested files into \code{dest_dir}.
\item If \code{convert_adc = TRUE}, modifies ADC files in place by inserting four empty columns after column 7.
\item Displays messages indicating the download status.
}
}
\description{
This function downloads specified IFCB data files from a given IFCB Dashboard URL.
It supports optional filename conversion and ADC file adjustments from the old IFCB file format.
}
\examples{
\dontrun{
ifcb_download_dashboard_data(
  dashboard_url = "https://ifcb-data.whoi.edu",
  samples = "IFCB1_2014_188_222013",
  file_types = c("blobs", "features"),
  dest_dir = "data",
  convert_filenames = FALSE,
  convert_adc = FALSE
)
}

}
