% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_download_test_data.R
\name{ifcb_download_test_data}
\alias{ifcb_download_test_data}
\title{Download test IFCB data}
\usage{
ifcb_download_test_data(dest_dir, method = "auto")
}
\arguments{
\item{dest_dir}{The destination directory where the files will be unzipped.}

\item{method}{Method to be used for downloading files. Current download methods are "internal",
"libcurl", "wget", "curl" and "wininet" (Windows only), and there is a value "auto":
see ‘utils::download.file’.}
}
\description{
This function downloads two zip archives containing .png images and MATLAB files from the `svea_skagerrak_kattegat`
dataset available in the SMHI IFCB plankton image reference library (Torstensson et al. 2024), and unzips them into the
specified folder. These data can, for instance, be used for testing `iRfcb`.
}
\examples{
\dontrun{
# Download and unzip IFCB test data into the "data" directory
ifcb_download_test_data("data")
}

}
\references{
Torstensson, Anders; Skjevik, Ann-Turi; Mohlin, Malin; Karlberg, Maria; Karlson, Bengt (2024). SMHI IFCB plankton image reference library. SciLifeLab. Dataset.
\doi{10.17044/scilifelab.25883455.v2}
}
