% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_extract_annotated_images.R
\name{ifcb_extract_annotated_images}
\alias{ifcb_extract_annotated_images}
\title{Extract Annotated Images from IFCB Data}
\usage{
ifcb_extract_annotated_images(
  manualdir,
  class2use_file,
  roidir,
  outdir,
  skip_class = NA
)
}
\arguments{
\item{manualdir}{A character string specifying the path to the directory containing the manually classified .mat files.}

\item{class2use_file}{A character string specifying the path to the file containing class names.}

\item{roidir}{A character string specifying the path to the directory containing the ROI files.}

\item{outdir}{A character string specifying the output directory where the extracted images will be saved.}

\item{skip_class}{A numeric value or vector specifying the class(es) to be skipped during the extraction process, e.g. unclassified. Default is NA.}
}
\value{
None. The function saves the extracted PNG images to the specified output directory.
}
\description{
This function extracts labelled images from IFCB (Imaging FlowCytobot) data,
annotated using the MATLAB code from the ifcb-analysis repository.
It reads manually classified data, maps class indices to class names, and extracts
the corresponding Region of Interest (ROI) images, saving them to the specified directory.
}
\examples{
\dontrun{
ifcb_extract_annotated_images(
  manualdir = "path/to/manualdir",
  class2use_file = "path/to/class2use_file.mat",
  roidir = "path/to/roidir",
  outdir = "path/to/outdir",
  skip_class = 1
)
}
}
\seealso{
\code{\link{ifcb_extract_pngs}} \code{\link{ifcb_extract_classified_images}} \url{https://github.com/hsosik/ifcb-analysis}
}
