% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_extract_biovolumes.R
\name{ifcb_extract_biovolumes}
\alias{ifcb_extract_biovolumes}
\title{Extract Biovolumes from IFCB Data and Compute Carbon Content}
\usage{
ifcb_extract_biovolumes(
  feature_files,
  mat_folder,
  class2use_file = NULL,
  micron_factor = 1/3.4,
  diatom_class = "Bacillariophyceae",
  marine_only = FALSE,
  threshold = "opt",
  multiblob = FALSE,
  feature_recursive = TRUE,
  mat_recursive = TRUE,
  use_python = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{feature_files}{A path to a folder containing feature files or a character vector of file paths.}

\item{mat_folder}{Path to the folder containing class or manual files.}

\item{class2use_file}{A character string specifying the path to the file containing the class2use variable (default NULL).}

\item{micron_factor}{Conversion factor for biovolume to cubic microns. Default is 1 / 3.4.}

\item{diatom_class}{A string vector of diatom class names in the World Register of Marine Species (WoRMS). Default is "Bacillariophyceae".}

\item{marine_only}{Logical. If TRUE, restricts the WoRMS search to marine taxa only. Default is FALSE.}

\item{threshold}{Threshold for selecting class information ("opt" or other, default is "opt").}

\item{multiblob}{A logical indicating whether to include multiblob features. Default is FALSE.}

\item{feature_recursive}{Logical. If TRUE, the function will search for feature files recursively within when \code{feature_files} is a folder. Default is TRUE.}

\item{mat_recursive}{Logical. If TRUE, the function will search for MATLAB files recursively within the \code{mat_folder}. Default is TRUE.}

\item{use_python}{Logical. If \code{TRUE}, attempts to read the \code{.mat} file using a Python-based method. Default is \code{FALSE}.}

\item{verbose}{A logical indicating whether to print progress messages. Default is TRUE.}
}
\value{
A data frame containing 'sample', 'classifier' 'roi_number', 'class', 'biovolume_um3', and computed 'carbon_pg'.
}
\description{
This function reads biovolume data from feature files generated by the \code{ifcb-analysis} repository (Sosik and Olson 2007)
and matches them with corresponding classification results or manual annotations. It calculates biovolume in cubic micrometer and
determines if each class is a diatom based World Register of Marine Species (WoRMS). Carbon content
is computed for each roi using specific conversion functions depending on whether the class
is identified as a diatom or not, according to Menden-Deuer and Lessard 2000.
}
\details{
The function combines biovolume data extracted from feature files with class
information read from corresponding MATLAB files. It determines if each class is a
diatom based on user-defined criteria and computes carbon content using conversion
functions specific to diatoms and non-diatom protists.

If \code{use_python = TRUE}, the function tries to read the \code{.mat} file using \code{ifcb_read_mat()}, which relies on \code{SciPy}.
This approach may be faster than the default approach using \code{R.matlab::readMat()}, especially for large \code{.mat} files.
To enable this functionality, ensure Python is properly configured with the required dependencies.
You can initialize the Python environment and install necessary packages using \code{ifcb_py_install()}.

If \code{use_python = FALSE} or if \code{SciPy} is not available, the function falls back to using \code{R.matlab::readMat()}.
}
\examples{
\dontrun{
# Example usage:
feature_files <- "data/features"
mat_folder <- "data/classified"
biovolume_df <- ifcb_extract_biovolumes(feature_files, mat_folder)
print(biovolume_df)
}

}
\references{
Menden-Deuer Susanne, Lessard Evelyn J., (2000), Carbon to volume relationships for dinoflagellates, diatoms, and other protist plankton, Limnology and Oceanography, 3, doi: 10.4319/lo.2000.45.3.0569.

Sosik, H. M. and Olson, R. J. (2007), Automated taxonomic classification of phytoplankton sampled with imaging-in-flow cytometry. Limnol. Oceanogr: Methods 5, 204–216.
}
\seealso{
\code{\link{ifcb_read_features}} \code{\link{ifcb_is_diatom}} \url{https://www.marinespecies.org/}
}
