% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_extract_biovolumes.R
\name{ifcb_extract_biovolumes}
\alias{ifcb_extract_biovolumes}
\title{Extract Biovolumes from IFCB Data and Compute Carbon Content}
\usage{
ifcb_extract_biovolumes(
  feature_files,
  mat_folder,
  class2use_file = NULL,
  micron_factor = 1/3.4,
  diatom_class = "Bacillariophyceae",
  threshold = "opt",
  multiblob = FALSE,
  feature_folder = deprecated(),
  class_folder = deprecated()
)
}
\arguments{
\item{feature_files}{A path to a folder containing feature files or a character vector of file paths.}

\item{mat_folder}{Path to the folder containing class or manual files.}

\item{class2use_file}{A character string specifying the path to the file containing the class2use variable (default NULL).}

\item{micron_factor}{Conversion factor for biovolume to cubic microns. Default is 1 / 3.4.}

\item{diatom_class}{A string vector of diatom class names in the World Register of Marine Species (WoRMS). Default is "Bacillariophyceae".}

\item{threshold}{Threshold for selecting class information ("opt" or other, default is "opt").}

\item{multiblob}{A logical indicating whether to include multiblob features. Default is FALSE.}

\item{feature_folder}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use \code{feature_files} instead.}

\item{class_folder}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Use \code{mat_folder} instead.}
}
\value{
A data frame containing 'sample', 'classifier' 'roi_number', 'class', 'biovolume_um3', and computed 'carbon_pg'.
}
\description{
This function reads biovolume data from feature files generated by the \code{ifcb-analysis} repository (Sosik and Olson 2007)
and matches them with corresponding classification results or manual annotations. It calculates biovolume in cubic micrometer and
determines if each class is a diatom based World Register of Marine Species (WoRMS). Carbon content
is computed for each roi using specific conversion functions depending on whether the class
is identified as a diatom or not, according to Menden-Deuer and Lessard 2000.
}
\details{
The function combines biovolume data extracted from feature files with class
information read from corresponding MATLAB files. It determines if each class is a
diatom based on user-defined criteria and computes carbon content using conversion
functions specific to diatoms and non-diatom protists.
}
\examples{
\dontrun{
# Example usage:
feature_files <- "data/features"
mat_folder <- "data/classified"
biovolume_df <- ifcb_extract_biovolumes(feature_files, mat_folder)
print(biovolume_df)
}

}
\references{
Menden-Deuer Susanne, Lessard Evelyn J., (2000), Carbon to volume relationships for dinoflagellates, diatoms, and other protist plankton, Limnology and Oceanography, 3, doi: 10.4319/lo.2000.45.3.0569.

Sosik, H. M. and Olson, R. J. (2007), Automated taxonomic classification of phytoplankton sampled with imaging-in-flow cytometry. Limnol. Oceanogr: Methods 5, 204–216.
}
\seealso{
\code{\link{ifcb_read_features}} \code{\link{ifcb_is_diatom}} \url{https://www.marinespecies.org/}
}
