% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_extract_classified_images.R
\name{ifcb_extract_classified_images}
\alias{ifcb_extract_classified_images}
\title{Extract Taxa Images from MATLAB Classified Sample}
\usage{
ifcb_extract_classified_images(
  sample,
  classifieddir,
  roidir,
  outdir,
  taxa = "All",
  threshold = "opt"
)
}
\arguments{
\item{sample}{A character string specifying the sample name.}

\item{classifieddir}{A character string specifying the directory containing the classified files.}

\item{roidir}{A character string specifying the directory containing the ROI files.}

\item{outdir}{A character string specifying the directory to save the extracted images.}

\item{taxa}{A character string specifying the taxa to extract. Default is "All".}

\item{threshold}{A character string specifying the threshold to use ("none", "opt", "adhoc"). Default is "opt".}
}
\value{
No return value, called for side effects. Extracts and saves taxa images to a directory.
}
\description{
This function reads a MATLAB classified sample file (.mat) generated
by the `start_classify_batch_user_training` function from the ifcb-analysis repository,
extracts specified taxa images from the corresponding ROI files,
and saves each image in a specified directory.
}
\examples{
\dontrun{
# Define the parameters
sample <- "D20230311T092911"
classifieddir <- "path/to/classifieddir"
roidir <- "path/to/roidir"
outdir <- "path/to/outputdir"
taxa <- "All"  # or specify a particular taxa
threshold <- "opt"  # or specify another threshold

# Extract taxa images from the classified sample
ifcb_extract_classified_images(sample, classifieddir, roidir, outdir, taxa, threshold)
}
}
\seealso{
\code{\link{ifcb_extract_pngs}} \code{\link{ifcb_extract_annotated_images}} \url{https://github.com/hsosik/ifcb-analysis}
}
