% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_get_ecotaxa_example.R
\name{ifcb_get_ecotaxa_example}
\alias{ifcb_get_ecotaxa_example}
\title{Get Ecotaxa Column Names}
\usage{
ifcb_get_ecotaxa_example()
}
\value{
A data frame containing Ecotaxa example metadata.
}
\description{
This function reads an example Ecotaxa metadata file included in the \code{iRfcb} package.
The example file originates from \url{https://github.com/VirginieSonnet/IFCBdatabaseToEcotaxa}
These columns are used for submitting IFCB data to \url{https://ecotaxa.obs-vlfr.fr/}.
}
\examples{
\dontrun{
ecotaxa_example <- ifcb_get_ecotaxa_example()
print(ecotaxa_example)
}
}
