% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_get_ferrybox_data.R
\name{ifcb_get_ferrybox_data}
\alias{ifcb_get_ferrybox_data}
\title{Get Ferrybox Data Based on Timestamps}
\usage{
ifcb_get_ferrybox_data(
  timestamps,
  ferrybox_folder,
  parameters = c("8002", "8003"),
  ship = "SveaFB"
)
}
\arguments{
\item{timestamps}{A vector of POSIXct timestamps for which GPS coordinates and parameter data are to be retrieved.}

\item{ferrybox_folder}{Path to the folder containing Ferrybox \code{.txt} files.}

\item{parameters}{String vector with parameters to extract from the Ferrybox data. Defaults to \code{c("8002", "8003")}.}

\item{ship}{Name of the ship to filter Ferrybox files (default is "SveaFB").}
}
\value{
A data frame containing the input timestamps and corresponding data for the specified parameters. Columns include 'timestamp' and the specified parameters.
}
\description{
This internal SMHI function reads \code{.txt} files from a specified folder containing Ferrybox data,
filters them based on a specified ship name (default is "SveaFB" for R/V Svea), and extracts
data (including GPS coordinates) for timestamps (rounded to the nearest minute) falling within the date ranges defined in the file names.
}
\examples{
\dontrun{
ferrybox_folder <- "/path/to/ferrybox/data"
timestamps <- as.POSIXct(c("2016-08-10 10:47:34 UTC",
                           "2016-08-10 11:12:21 UTC",
                           "2016-08-10 11:35:59 UTC"))

result <- ifcb_get_ferrybox_data(timestamps, ferrybox_folder)
print(result)
}

}
