% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_get_shark_colnames.R
\name{ifcb_get_shark_colnames}
\alias{ifcb_get_shark_colnames}
\title{Get Shark Column Names}
\usage{
ifcb_get_shark_colnames()
}
\value{
A data frame containing the SHARK column names.
}
\description{
This function reads SHARK column names from a specified tab-separated values (TSV) file included in the package.
These columns are used for submitting IFCB data to \url{https://sharkweb.smhi.se/}.
}
\examples{
shark_colnames <- ifcb_get_shark_colnames()
print(shark_colnames)
}
\seealso{
\code{\link{ifcb_get_shark_example}}
}
