% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_get_shark_example.R
\name{ifcb_get_shark_example}
\alias{ifcb_get_shark_example}
\title{Get Shark Column Example}
\usage{
ifcb_get_shark_example()
}
\value{
A data frame containing example data following the SHARK submission format.
}
\description{
This function reads a SHARK submission example from a file included in the package.
This format is used for submitting IFCB data to \url{https://shark.smhi.se/en/}.
}
\examples{
shark_example <- ifcb_get_shark_example()

# Print example as tibble
print(shark_example)
}
\seealso{
\code{\link{ifcb_get_shark_colnames}}
}
