% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_is_diatom.R
\name{ifcb_is_diatom}
\alias{ifcb_is_diatom}
\title{Identify Diatoms in Taxa List}
\usage{
ifcb_is_diatom(taxa_list, diatom_class = "Bacillariophyceae")
}
\arguments{
\item{taxa_list}{A character vector containing the list of taxa names.}

\item{diatom_class}{A character string specifying the class name to be identified as diatoms.
Default is "Bacillariophyceae".}
}
\value{
A logical vector indicating whether each cleaned taxa name belongs to the specified diatom class.
}
\description{
This function takes a list of taxa names, cleans them, retrieves their
corresponding classification records from the World Register of Marine Species (WoRMS),
and checks if they belong to the specified diatom class. The function only uses the
first name (genus name) of each taxa for classification.
}
\examples{
taxa_list <- c("Nitzschia_sp", "Chaetoceros_sp", "Dinophysis_norvegica", "Thalassiosira_sp")
ifcb_is_diatom(taxa_list)

}
\seealso{
\url{https://www.marinespecies.org/}
}
