% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_py_install.R
\name{ifcb_py_install}
\alias{ifcb_py_install}
\title{Install iRfcb Python Environment}
\usage{
ifcb_py_install(..., envname = ".virtualenvs/iRfcb")
}
\arguments{
\item{...}{Additional arguments passed to \code{virtualenv_create}, such as \code{packages}.}

\item{envname}{A character string specifying the name of the virtual environment to create. Default is "~/.virtualenvs/iRfcb".}
}
\description{
This function creates and activates a Python virtual environment named "iRfcb" and installs the required Python packages as specified in the "requirements.txt" file.
Additional packages can be installed by passing them through the \code{packages} argument.
}
\examples{
\dontrun{
# Install the iRfcb Python environment
ifcb_py_install()

# Install the iRfcb Python environment with additional packages
ifcb_py_install(packages = c("numpy", "pandas"))
}
}
