% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{ifcb_py_install}
\alias{ifcb_py_install}
\title{Install iRfcb Python Environment}
\usage{
ifcb_py_install(..., envname = "/.virtualenvs/iRfcb")
}
\arguments{
\item{...}{Additional arguments passed to `virtualenv_create`.}

\item{envname}{A character string specifying the name of the virtual environment to create. Default is "iRfcb".}
}
\description{
This function creates a Python virtual environment named "iRfcb" and installs the required Python packages as specified in the "requirements.txt" file.
}
\examples{
\dontrun{
# Install the iRfcb Python environment
ifcb_install_iRfcb()
}
}
