% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_read_features.R
\name{ifcb_read_features}
\alias{ifcb_read_features}
\title{Read feature files from a specified folder}
\usage{
ifcb_read_features(feature_folder, multiblob = FALSE)
}
\arguments{
\item{feature_folder}{Path to the folder containing feature files.}

\item{multiblob}{Logical indicating whether to filter for multiblob files (default: FALSE).}
}
\value{
A named list of data frames, where each element corresponds to a feature file read from \code{feature_folder}.
  The list is named with the base names of the feature files.
}
\description{
This function reads feature files from a given folder, filtering them optionally
based on whether they are multiblob or single blob files.
}
\examples{
# Read feature files from a folder
features <- ifcb_read_features("path/to/feature_folder")

# Read only multiblob feature files
multiblob_features <- ifcb_read_features("path/to/feature_folder", multiblob = TRUE)

}
