% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_read_mat.R
\name{ifcb_read_mat}
\alias{ifcb_read_mat}
\title{Read a MATLAB .mat File in R}
\usage{
ifcb_read_mat(file_path)
}
\arguments{
\item{file_path}{A character string representing the full path to the .mat file.}
}
\value{
A list containing the MATLAB variables.
}
\description{
This function reads a MATLAB .mat file using a Python function via \code{reticulate}.
}
\details{
This function requires a Python interpreter with \code{SciPy} installed.

This function requires a python interpreter to be installed.
The required python packages can be installed in a virtual environment using \code{ifcb_py_install}.
}
\examples{
\dontrun{
data <- read_mat_file_r("C:/data/sample.mat")
}

}
\seealso{
\code{\link{ifcb_py_install}}
}
