% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IFCB_volume_analyzed.R
\name{IFCB_volume_analyzed}
\alias{IFCB_volume_analyzed}
\title{Estimate Volume Analyzed from IFCB Header File}
\usage{
IFCB_volume_analyzed(hdrfilename, hdrOnly_flag = 0)
}
\arguments{
\item{hdrfilename}{A character vector specifying the path(s) to one or more .hdr files or URLs.}

\item{hdrOnly_flag}{An optional flag indicating whether to skip ADC file estimation (default is 0).}
}
\value{
A numeric vector containing the estimated sample volume analyzed for each header file.
}
\description{
This function reads an IFCB header file to extract sample run time and inhibittime,
and returns the associated estimate of sample volume analyzed (in milliliters).
The function assumes a standard IFCB configuration with a sample syringe operating
at 0.25 mL per minute. For IFCB instruments after 007 and higher (except 008).
}
\examples{
\dontrun{
# Example: Estimate volume analyzed from an IFCB header file
hdr_file <- "path/to/IFCB_hdr_file.hdr"
ml_analyzed <- IFCB_volume_analyzed(hdr_file)
print(ml_analyzed)
}
}
