% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRfcb-package.R
\docType{package}
\name{iRfcb-package}
\alias{iRfcb}
\alias{iRfcb-package}
\title{iRfcb: Tools for Analyzing and Processing Imaging FlowCytobot (IFCB) Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive suite of tools for managing, processing, and analyzing data from the Imaging FlowCytobot (IFCB). ‘I R FlowCytobot’ (iRfcb) supports quality control, geospatial analysis, and preparation of IFCB data for publication in databases like \url{https://www.gbif.org}, \url{https://www.obis.org}, \url{https://emodnet.ec.europa.eu/en}, \url{https://shark.smhi.se/}, and \url{https://www.ecotaxa.org}. The package integrates with the MATLAB ifcb-analysis tool and offers features for working with annotated, classified, and manually classified image datasets. Key functionalities include image extraction, particle size distribution analysis, taxonomic data handling, and biomass concentration calculations, essential for plankton research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://europeanifcbgroup.github.io/iRfcb/}
  \item \url{https://github.com/EuropeanIFCBGroup/iRfcb}
}

}
\author{
\strong{Maintainer}: Anders Torstensson \email{anders.torstensson@smhi.se} (\href{https://orcid.org/0000-0002-8283-656X}{ORCID}) (Swedish Meteorological and Hydrological Institute)

}
\keyword{internal}
