% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRfcb-package.R
\docType{package}
\name{iRfcb-package}
\alias{iRfcb}
\alias{iRfcb-package}
\title{iRfcb: I 'R' FlowCytobot (iRfcb): Tools for Analyzing and Processing Data from the IFCB}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive suite of tools for managing and analyzing data from the Imaging FlowCytobot (IFCB).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://europeanifcbgroup.github.io/iRfcb/}
  \item \url{https://github.com/EuropeanIFCBGroup/iRfcb}
}

}
\author{
\strong{Maintainer}: Anders Torstensson \email{anders.torstensson@smhi.se} (\href{https://orcid.org/0000-0002-8283-656X}{ORCID}) (Swedish Meteorological and Hydrological Institute)

}
\keyword{internal}
