% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifcb_helper_functions.R
\name{read_hdr_file}
\alias{read_hdr_file}
\title{Function to Read Individual Files and Extract Relevant Lines}
\usage{
read_hdr_file(file)
}
\arguments{
\item{file}{A character string specifying the path to the HDR file.}
}
\value{
A data frame with columns: \code{parameter}, \code{value}, and \code{file}.
}
\description{
This function reads an HDR file and extracts relevant lines containing parameters and their values.
}
